[libc++] Refactor the tests for std::random_device
That will make it easier to change the behavior of the arc4random()
based implementation. Note that in particular, the eval.pass.cpp test
used to work with non "/dev/random" based implementations because we'd
throw an exception upon constructing the random_device. This patch makes
the intent of the test clearer.

NOKEYCHECK=True
GitOrigin-RevId: 84654f2733f63dc725a7b3d7c55d56849d2d9358